% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{download_sample_data}
\alias{download_sample_data}
\title{Download sample annotation data}
\usage{
download_sample_data(destination = ".")
}
\arguments{
\item{destination}{The directory where to store the data.}
}
\description{
Download sample annotation data
}
\examples{
\dontrun{
download_sample_data()
}
}
