% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{job_species}
\alias{job_species}
\title{View species in an annotation}
\usage{
job_species(annotation_id)
}
\arguments{
\item{annotation_id}{The annotation id from from my_jobs().}
}
\value{
A data.frame, with species and counts in this annotation.
}
\description{
Requires login, please run mol_login(<email_address>)
Uses the output from my_jobs().
}
\examples{
\dontrun{
job_species(<annotation_id>)
}
}
