% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melopsittacus_short}
\alias{melopsittacus_short}
\title{Short budgerigar sample dataset}
\format{
A data frame with 1500 rows and 10 variables:
}
\source{
See introduction vignette
}
\usage{
melopsittacus_short
}
\description{
Dataset containing 1500 annotated records of the budgerigar. Subsampled
version of the dataset downloaded using download_sample_data().
}
\keyword{datasets}
