% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{mol_login}
\alias{mol_login}
\title{Map of Life Login}
\usage{
mol_login(email, password = NULL)
}
\arguments{
\item{email}{The email address associated with your Map of Life Account.}

\item{password}{Your map of life password. If left blank, and you are in RStudio you can enter it via a secure popup.}
}
\value{
No return value
}
\description{
Login to your Map of Life account.
}
\examples{
\dontrun{
mol_login("your.email@company.com")
}
}
