% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{my_jobs}
\alias{my_jobs}
\title{List logged in users current annotation jobs.}
\usage{
my_jobs()
}
\value{
A data.frame containing jobs metadata.
}
\description{
Requires login, please run mol_login(<email_address>)
}
\examples{
\dontrun{
my_jobs()
}
}
