% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ninox_short}
\alias{ninox_short}
\title{Short powerful owl sample dataset}
\format{
A data frame with 1500 rows and 11 variables:
}
\source{
See introduction vignette
}
\usage{
ninox_short
}
\description{
Dataset containing 1500 annotated records of the powerful owl. Subsampled
version of the dataset downloaded using download_sample_data().
}
\keyword{datasets}
