% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{read_output}
\alias{read_output}
\title{Read annotation output into R}
\usage{
read_output(directory, drop_event_id = TRUE)
}
\arguments{
\item{directory}{The path of the data.}

\item{drop_event_id}{Whether to drop the event_id column or not.}
}
\value{
A data.frame of annotated data, one row per variable per event
}
\description{
Convenience function which reads and joins annotation results spread across multiple files for space efficiency.
To run this function, please first download annotated data first using download_annotation()
}
\examples{
\dontrun{
read_output("path/to/your/downloaded/data/directory")
}
}
