% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{start_annotation_simple}
\alias{start_annotation_simple}
\title{Submit a dataframe for on the fly annotation}
\usage{
start_annotation_simple(events, layers)
}
\arguments{
\item{events}{A data.frame with the following columns: event_id (character), lat (numeric), lng (numeric), date (string/date) in the format YYYY-MM-DD}

\item{layers}{A list of parameters or vector of codes, of the layers, see the examples below.}
}
\description{
Does not require login - for use for small numbers of records and pilot jobs
}
\examples{
\dontrun{
events <- data.frame(
   event_id = as.character(1:2),
   lng = c(-4, 24),
   lat = c(10, 10),
   date = '2015-01-01'
)
layers <- 'landsat8-evi-100-16'
start_annotation_simple(events, layers)
}
}
