% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{download_annotation}
\alias{download_annotation}
\title{Download annotation results}
\usage{
download_annotation(annotation_id, dir = "annotation_results")
}
\arguments{
\item{annotation_id}{The id of the annotation}

\item{dir}{The directory where to write the annotation.}
}
\value{
The path of the downloaded annotation.
}
\description{
Download results of a successfully completed batch annotation.
Requires login, please run mol_login(<email_address>)
Uses the output from my_jobs() for the annotation id.
}
\examples{
\dontrun{
download_annotation(<annotation_id>, <dir>)
}
}
