% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{get_products}
\alias{get_products}
\title{Retrieve product metadata}
\usage{
get_products()
}
\value{
A data.frame of spatial and temporal buffer limits
}
\description{
Get information on available products for annotation.
Get the spatial and temporal buffer limits for use in when creating a custom annotation.
}
\examples{
\dontrun{
get_products()
}
}
