% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{job_details}
\alias{job_details}
\title{Retrieve annotation job details}
\usage{
job_details(annotation_id)
}
\arguments{
\item{annotation_id}{The annotation id from from my_jobs().}
}
\value{
A data.frame of layers and their statuses, along with the annotation_id, and the dataset_id for the custom annotation.
}
\description{
Get details of a batch annotation job. Requires login, please run mol_login(<email_address>).
Uses the output from my_jobs().
}
\examples{
\dontrun{
job_details(<annotation_id>)
}
}
