% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{my_datasets}
\alias{my_datasets}
\title{View datasets}
\usage{
my_datasets()
}
\value{
A data frame of a users datasets and their associated ids.
}
\description{
List logged-in user's uploaded datasets (uploaded to https://mol.org through https://mol.org/upload).
Requires login, please run mol_login(<email_address>)
}
\examples{
\dontrun{
my_datasets()
}
}
