% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{my_jobs}
\alias{my_jobs}
\title{List all jobs}
\usage{
my_jobs()
}
\value{
A data.frame containing jobs metadata.
}
\description{
List logged-in user's past and current annotation jobs.
Requires login, please run mol_login(<email_address>)
}
\examples{
\dontrun{
my_jobs()
}
}
